refactor(acp): isolate protocol facade and stream observation
